phonenumber/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
phonenumbers==8.12.29
googlesearch-python==1.0.1